Remove bot method StopReceivingUpdates()
The StopReceivingUpdates() was used to stop a goroutine created at
GetUpdatesChannel().

If StopReceivingUpdates() were called twice or if GetUpdatesChannel()
were called after a StopReceivingUpdates() the bot behavior would be
incorrect due to 'shutdownChannel' close.

This commit removes StopReceivingUpdates() and modifies UpdatesChannel
struct to have a Shutdown() method which can stop both the associated
goroutine and http handler to feed the updates channel.

Signed-off-by: Vinicius Tinti <viniciustinti@gmail.com>