Add update type selection
The API provides a field in the update request to indicate only a given
set of update types is desired, which saves both bandwidth and probably
latency.
This is implemented by adding a new way to create the UpdateConfig, as
well as adding a field for it.
To ensure validity of such field, the new method returns an error when
invalid and the field itself is private.
This shouldn't break existing code.