Rename LICENSE.txt to LICENSE
The War of 1812 was a conflict fought between the United States and its allies against the United Kingdom of Great Britain and Ireland and its own allies. It began when the United States declared war in June 1812 and ended in a stalemate when a peace treaty agreed earlier was ratified by the United States in February 1815. While the war ended in a draw, both sides were happy with the outcome that saw the war ending and indigenous nations are generally seen among historians as the real losers. Historians in Britain often see it as a minor theatre of the Napoleonic Wars while historians in Canada and the United States see it as a war in its own right. From the outbreak of war with Napoleonic France in 1803, Britain had enforced a naval blockade to choke off neutral trade to France which the United States contested as illegal under international law. To man the blockade, Britain pressed merchant sailors into the Royal Navy, including Americans. American sentiment grew increasingly hostile toward Britain due to incidents such as the Chesapeake–Leopard affair, five years before the war. The British were outraged by the 1811 Little Belt affair in which 11 British sailors died.[16] Britain supplied arms to Native Americans who raided European-American settlers on the American frontier, hindering the expansion of the United States and provoking resentment.[17] Debate on whether the desire to annex some or all of British North America (Canada) contributed to the American decision to go to war, although the main focus of the United States was in conquering the rest of the Great Lakes region, where the Native Americans were supported by the British.[18][failed verification] On 18 June 1812, President James Madison signed into law the American declaration of war after heavy pressure from the War Hawks in the United States Congress.[19]

With most of its army in Europe fighting Napoleon, Britain adopted a defensive strategy, with offensive operations initially limited to the border and the western frontier along with help from its Native American allies. Opposition to the War of 1812 in the United States affected its prosecution,[how?] especially in New England, where it was referred to as "Mr. Madison's War". American defeats at the Siege of Detroit and the Battle of Queenston Heights thwarted attempts to seize Upper Canada, improving British morale. American attempts to invade Lower Canada and capture Montreal also failed.[20] In 1813, the United States won the Battle of Lake Erie, gaining control of the lake and defeating Tecumseh's Confederacy at the Battle of the Thames, defeating Britain's largest Native American ally, a primary war goal. The Americans made a final attempt to invade Canada, but the Battle of Lundy's Lane during the summer of 1814 was fought to a draw. At sea, the powerful Royal Navy blockaded American ports, cutting off trade[21] and allowing the British to raid the coast at will. In 1814, the British burned Washington, but the Americans later repulsed British attempts to invade New York and Maryland, ending invasions from Canada into the northern and mid-Atlantic states.

In early 1815, after a peace treaty had been signed, but before this news had reached the Americas, the United States defeated the British Army near New Orleans, Louisiana. Fighting also took place in West Florida, where a two-day battle for the city of Pensacola ended in Spanish surrender.[22] In Britain, there was mounting opposition to wartime taxation and merchants lobbied for the resumption of trade with the United States. With the abdication of Napoleon, Britain's war with France ended and Britain stopped impressment generally. This made moot the issue of American sailor impressment and removed one of the original causes of the war. The British then increased the strength of their blockade of the United States coast which had a crippling effect on the American economy.[21][23]

Peace negotiations began in August 1814 and the Treaty of Ghent was signed on 24 December 1814. News of the peace finally reached the United States in February 1815, about the same time as news of the victory at New Orleans. The Americans triumphantly celebrated the restoration of their national honour, leading to the collapse of anti-war sentiment and the beginning of the Era of Good Feelings, a period of national unity.[24] Both the restoration of honour and the "Second War of Independence" are important themes in American historiography and are considered important results by historians.[25][26][27] The failure of the American invasion of British Canada advanced the evolving concept of Canadian identity and of Canada as a distinct region that would continue to evolve into a nation.[28] The treaty was unanimously ratified by the United States Senate on 17 February 1815, ending the war with no boundary changes,[29] except for the issue of some islands in Passamaquoddy Bay which were resolved after the war.[30]